+++ docs/integrations/metadata-stores/databases/clickhouse.md
@@ -5,7 +5,7 @@ description: "Learn how to use ClickHouse as a Metaxy metadata store."
 
 # Metaxy + ClickHouse
 
-[ClickHouse](https://clickhouse.com/) is a (1) column-oriented OLAP database designed for real-time analytics. To use Metaxy with ClickHouse, configure [`ClickHouseMetadataStore`][metaxy.ext.metadata_stores.clickhouse.ClickHouseMetadataStore]. Versioning computations run natively in ClickHouse, making it well-suited for high-throughput production workloads.
+[ClickHouse](https://clickhouse.com/) is a (1) column-oriented OLAP database designed for real-time analytics. To use Metaxy with ClickHouse, configure [`ClickHouseMetadataStore`][metaxy.ext.clickhouse.ClickHouseMetadataStore]. Versioning computations run natively in ClickHouse, making it well-suited for high-throughput production workloads.
 { .annotate }
 
 1. extremely fast
@@ -90,13 +90,13 @@ pip install clickhouse-sqlalchemy
 ## API Reference
 
 <!-- dprint-ignore-start -->
-::: metaxy.ext.metadata_stores.clickhouse
+::: metaxy.ext.clickhouse
     options:
       members: false
       show_root_heading: true
       heading_level: 2
 
-::: metaxy.ext.metadata_stores.clickhouse.ClickHouseMetadataStore
+::: metaxy.ext.clickhouse.ClickHouseMetadataStore
     options:
       members: false
       heading_level: 3
@@ -106,7 +106,7 @@ pip install clickhouse-sqlalchemy
 
 <!-- dprint-ignore-start -->
 ::: metaxy-config
-    class: metaxy.ext.metadata_stores.clickhouse.ClickHouseMetadataStoreConfig
+    class: metaxy.ext.clickhouse.ClickHouseMetadataStoreConfig
     path_prefix: stores.dev.config
     header_level: 3
 <!-- dprint-ignore-end -->
