+++ examples/example-aggregation/metaxy.toml
@@ -3,7 +3,7 @@ entrypoints = ["example_aggregation.features"]
 auto_create_tables = true # Enable for development/examples
 
 [stores.dev]
-type = "metaxy.ext.metadata_stores.delta.DeltaMetadataStore"
+type = "metaxy.ext.polars.DeltaMetadataStore"
 
 [stores.dev.config]
 root_path = "/tmp/example_aggregation"
