+++ examples/example-ducklake/metaxy.toml
@@ -3,7 +3,7 @@ entrypoints = ["example_ducklake.definitions"]
 auto_create_tables = true # Enable for development/examples
 
 [stores.dev]
-type = "metaxy.ext.metadata_stores.duckdb.DuckDBMetadataStore"
+type = "metaxy.ext.duckdb.DuckDBMetadataStore"
 
 [stores.dev.config]
 database = "${DUCKLAKE_DEMO_DB:-:memory:}"
