+++ examples/example-ducklake/src/example_ducklake/pipeline.py
@@ -2,7 +2,7 @@
 
 import metaxy as mx
 import polars as pl
-from metaxy.ext.metadata_stores.duckdb import DuckDBMetadataStore
+from metaxy.ext.duckdb import DuckDBMetadataStore
 from metaxy.models.constants import METAXY_PROVENANCE_BY_FIELD
 
 from example_ducklake.definitions import DuckLakeDemoFeature
