+++ src/metaxy/ext/bigquery/__init__.py
@@ -1,0 +1,5 @@
+"""BigQuery metadata store extension."""
+
+from metaxy.ext.bigquery.metadata_store import BigQueryMetadataStore, BigQueryMetadataStoreConfig
+
+__all__ = ["BigQueryMetadataStore", "BigQueryMetadataStoreConfig"]
