+++ src/metaxy/ext/clickhouse/__init__.py
@@ -1,12 +1,9 @@
-"""Deprecated: use metaxy.ext.clickhouse.metadata_store instead."""
+"""ClickHouse metadata store extension."""
 
-from metaxy._warnings import _warn_deprecated_module
 from metaxy.ext.clickhouse.metadata_store import (
     ClickHouseMetadataStore,
     ClickHouseMetadataStoreConfig,
     ClickHouseVersioningEngine,
 )
-
-_warn_deprecated_module("metaxy.ext.metadata_stores.clickhouse", "metaxy.ext.clickhouse")
 
 __all__ = ["ClickHouseMetadataStore", "ClickHouseMetadataStoreConfig", "ClickHouseVersioningEngine"]
