+++ src/metaxy/ext/lancedb/__init__.py
@@ -1,0 +1,5 @@
+"""LanceDB metadata store extension."""
+
+from metaxy.ext.lancedb.metadata_store import LanceDBMetadataStore, LanceDBMetadataStoreConfig
+
+__all__ = ["LanceDBMetadataStore", "LanceDBMetadataStoreConfig"]
