+++ src/metaxy/ext/postgresql/__init__.py
@@ -1,0 +1,5 @@
+"""PostgreSQL metadata store extension."""
+
+from metaxy.ext.postgresql.metadata_store import PostgreSQLMetadataStore, PostgreSQLMetadataStoreConfig
+
+__all__ = ["PostgreSQLMetadataStore", "PostgreSQLMetadataStoreConfig"]
